requirements/dev.txt
-r commum.txt
coverage==7.4.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0